+++ tests/solver/broyden1_jac/polynom.rs
@@ -11,6 +11,7 @@ use nrf::test_cases::polynom;
 #[test]
 fn square() {
     let problem_size = 1;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         polynom::square2,
@@ -19,13 +20,14 @@ fn square() {
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
             UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
+        damping,
     );
 }
 
 #[test]
-#[should_panic]
 fn root_with_high_derivative() {
     let problem_size = 1;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         polynom::root_with_high_derivative,
@@ -34,12 +36,14 @@ fn root_with_high_derivative() {
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
             UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
+        damping,
     );
 }
 
 #[test]
 fn square_jac() {
     let problem_size = 1;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         polynom::square2,
@@ -49,13 +53,14 @@ fn square_jac() {
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
             UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
+        damping,
     );
 }
 
 #[test]
-#[should_panic]
 fn root_with_high_derivative_jac() {
     let problem_size = 1;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         polynom::root_with_high_derivative,
@@ -65,5 +70,6 @@ fn root_with_high_derivative_jac() {
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
             UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
+        damping,
     );
 }