+++ tests/solver/newton_raphson/broyden.rs
@@ -7,18 +7,21 @@ use crate::common::{run_test_case_fd, run_test_case_jac};
 #[test]
 fn broyden_case5_fd() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case5,
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case5_jac() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case5,
@@ -26,24 +29,28 @@ fn broyden_case5_jac() {
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case6_fd() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case6,
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case6_jac() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case6,
@@ -51,24 +58,28 @@ fn broyden_case6_jac() {
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case7_fd() {
     let problem_size = 10;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case7,
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case7_jac() {
     let problem_size = 10;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case7,
@@ -76,24 +87,28 @@ fn broyden_case7_jac() {
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case8_fd() {
     let problem_size = 20;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case8,
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case8_jac() {
     let problem_size = 20;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case8,
@@ -101,24 +116,28 @@ fn broyden_case8_jac() {
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case9_fd() {
     let problem_size = 2;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case9,
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }
 
 #[test]
 fn broyden_case9_jac() {
     let problem_size = 2;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case9,
@@ -126,6 +145,7 @@ fn broyden_case9_jac() {
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         nrf::solver::ResolutionMethod::NewtonRaphson,
+        damping,
     );
 }