+++ tests/solver/stationary_newton/broyden.rs
@@ -9,18 +9,21 @@ use crate::common::{run_test_case_fd, run_test_case_jac};
 #[test]
 fn broyden_case5_fd() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case5,
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case5_jac() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case5,
@@ -28,24 +31,28 @@ fn broyden_case5_jac() {
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case6_fd() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case6,
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case6_jac() {
     let problem_size = 5;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case6,
@@ -53,24 +60,28 @@ fn broyden_case6_jac() {
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case7_fd() {
     let problem_size = 10;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case7,
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case7_jac() {
     let problem_size = 10;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case7,
@@ -78,24 +89,28 @@ fn broyden_case7_jac() {
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case8_fd() {
     let problem_size = 20;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case8,
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case8_jac() {
     let problem_size = 20;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case8,
@@ -103,24 +118,28 @@ fn broyden_case8_jac() {
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case9_fd() {
     let problem_size = 2;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case9,
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }
 
 #[test]
 fn broyden_case9_jac() {
     let problem_size = 2;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case9,
@@ -128,6 +147,7 @@ fn broyden_case9_jac() {
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::StationaryNewton),
+        damping,
     );
 }